+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 1. Show fetching status at footer play bar. 在播放底栏显示加载状态。
 2. Use HarmonyOS-Sans as default font on Windows. 在 Windows 使用 HarmonyOS-Sans 作为默认字体。
+3. Optimized the layout of footer player on compact width. 优化底部播放器在窄屏下的布局。
 
 ## Fixes 修复